250404_11567_fixed

+++ Study_250407_BellmanFord_FloydWarshall/src/com/ssafy/minji/BOJ_11657_타임머신.java
@@ -5,10 +5,7 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.StringTokenizer;
 
 public class BOJ_11657_타임머신 {
@@ -20,8 +17,8 @@ public class BOJ_11657_타임머신 {
 
 		List<int[]> list = new ArrayList<>();
 
-		int[] dist = new int[N + 1];
-		Arrays.fill(dist, Integer.MAX_VALUE);
+		long[] dist = new long[N + 1];
+		Arrays.fill(dist, Long.MAX_VALUE);
 		dist[1] = 0;
 
 		int from;
@@ -40,7 +37,7 @@ public class BOJ_11657_타임머신 {
 				from = cur[0];
 				to = cur[1];
 				cost = cur[2];
-				if (dist[from] != Integer.MAX_VALUE && dist[to] > dist[from] + cost) {
+				if (dist[from] != Long.MAX_VALUE && dist[to] > dist[from] + cost) {
 					dist[to] = dist[from] + cost;
 				}
 			}
@@ -51,7 +48,7 @@ public class BOJ_11657_타임머신 {
 			from = cur[0];
 			to = cur[1];
 			cost = cur[2];
-			if (dist[from] != Integer.MAX_VALUE && dist[to] > dist[from] + cost) {
+			if (dist[from] != Long.MAX_VALUE && dist[to] > dist[from] + cost) {
 				cycleFlag = true;
 				System.out.println(-1);
 				return;
@@ -59,7 +56,7 @@ public class BOJ_11657_타임머신 {
 		}
 		
 		for(int i = 2 ; i < N + 1 ; i++) {
-			if(dist[i] == Integer.MAX_VALUE) {
+			if(dist[i] == Long.MAX_VALUE) {
 				System.out.println(-1);
 			}else {
 				System.out.println(dist[i]);
